+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Fix reverse `OneToOneField` relations not being persisted to the database when passed as kwargs to `baker.make()` ([#473](https://github.com/model-bakers/model_bakery/issues/473))
+
 ### Removed
 
 ## [1.23.4](https://pypi.org/project/model-bakery/1.23.4/)